pyproject.toml
[tool.poetry]
name = "project-5-habit-tracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
python-dotenv = "^1.1.1"
djangorestframework = "^3.16.0"
celery = "^5.5.3"
eventlet = "^0.40.1"
redis = "^6.2.0"
psycopg2-binary = "^2.9.10"
pillow = "^11.2.1"
djangorestframework-simplejwt = "^5.5.0"
coverage = "^7.9.1"
django-celery-beat = "^2.8.1"
requests = "^2.32.4"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
 exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py

)
'''

[tool.isort]
# максимальная длина строки
line_length = 119
